# W2MShip/src/main/java/com/example/w2mExample/dto/ShipDTO.java
package com.example.w2mExample.dto;

import com.fasterxml.jackson.annotation.JsonInclude;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class ShipDTO  {
	// input DTO. field 'private Integer id;' its used in ResponseDTO

	@NotNull
	@NotBlank
	@Size(max = 20)
	private String name;
}